fund-and-project-timelines: Create one graph per Fund
This is due to a problem where Plotly would resort tasks which
means that projects are not under the correct funds any more.

I couldn't find an option to change this and other options for
grouping did not seem to do what we want.

Also set option so that all labels on Y axis are shown